Support any number of configurable HTTP embedding endpoints
Generalizes the hash+one-HTTP-endpoint embedding system (PRs #57-60) to
any number of independently configured, independently enabled HTTP
embedding endpoints (each its own base URL/model/API key/dimensions/rate
limit) alongside the built-in hash provider -- the actual "add embeddings
from other models and endpoints" feature, not just a second hardcoded
provider.

- New domain.EmbeddingHTTPEndpoint + embedding_http_endpoints table +
  ports.EmbeddingEndpointStore, following the existing ScheduledCrawl
  CRUD precedent. document_embeddings.provider and the pgvector ANN
  column/index layer were already provider-string-keyed and needed no
  mechanical change -- only OperationalSettingsValues' flat HTTP fields,
  the two-value provider enum, bootstrap.NewEmbedders, and the admin UI
  actually assumed exactly {hash, http}.
- domain.ReconcileActiveProvider replaces OperationalSettings.Set's old
  static self-healing for EmbeddingProvider, since validity now depends
  on the dynamically configured endpoint list -- applied by
  bootstrap.SyncSettings on every poll tick and by the settings POST
  handler.
- Per-endpoint rate limiting (embedRateLimiter keyed by provider) replaces
  the old single shared EmbeddingRateLimitPerSecond, since endpoints can
  be genuinely different external services with different limits.
- Admin API: full CRUD at /admin/api/embeddings/endpoints(/{id}), with
  API-key encryption at rest (settingscrypto) and the established
  never-echo-a-secret-back convention; POST .../models and .../test
  generalized from "the one saved HTTP config" to a not-yet-saved
  candidate probe, so an endpoint can be tested before being created.
- Admin UI: new endpoint list + add/edit subpage (mirroring the
  scheduled_crawls list/edit-subpage pattern), Settings' embedding-provider
  selector now populated dynamically from live endpoints instead of a
  static two-option select.
- One-time migration (guarded by a persisted marker, not by table
  row-count, so a deliberately-deleted migrated endpoint never gets
  resurrected) carries an existing single-HTTP-endpoint install's live
  config into the new table, so upgrading se.mo-sys.de doesn't lose its
  configured IONOS setup.

Verified against a real Postgres+pgvector container (3 simultaneously
enabled providers of different dimensions, full endpoint CRUD, both
migration paths) in addition to the SQLite suite.

+++ internal/application/embed_rate_limiter.go
@@ -8,14 +8,14 @@ import (
 
 // embedRateLimiter paces calls to a shared ports.EmbeddingProvider so
 // their combined rate -- not just one caller's own -- respects a
-// configured requests-per-second cap (domain.OperationalSettingsValues.
-// EmbeddingRateLimitPerSecond). Safe for concurrent use: sqlCrawlerService
-// holds one instance shared across every crawl job running concurrently
-// on this process (up to maxConcurrentCrawls at once -- see
+// configured requests-per-second cap (domain.EmbeddingHTTPEndpoint.
+// RateLimitPerSecond). Safe for concurrent use: sqlCrawlerService holds one
+// instance per enabled provider, each shared across every crawl job running
+// concurrently on this process (up to maxConcurrentCrawls at once -- see
 // internal/adapters/restapi/crawl_internal.go), so their Embed calls
-// interleave at the shared rate instead of each job independently pacing
-// itself and letting the combined rate exceed the limit. The zero value
-// is ready to use.
+// against that one provider interleave at its shared rate instead of each
+// job independently pacing itself and letting the combined rate exceed the
+// limit. The zero value is ready to use.
 type embedRateLimiter struct {
 	mu   sync.Mutex
 	next time.Time
@@ -25,9 +25,8 @@ type embedRateLimiter struct {
 // returns early if ctx is cancelled first -- callers don't need to check
 // which happened: an early return here just means the actual Embed call
 // fails fast against the same cancelled context instead. ratePerSecond <=
-// 0 disables pacing entirely -- domain.OperationalSettings.Set never
-// actually produces that in production (it self-heals to a positive
-// default), but this lets a test opt out of real waits.
+// 0 disables pacing entirely -- a local provider with no rate limit of its
+// own (e.g. Ollama), or the built-in hash provider, passes 0 here.
 //
 // Each call atomically reserves the next available slot (now, or right
 // after whichever slot was most recently reserved, whichever is later)
@@ -35,7 +34,7 @@ type embedRateLimiter struct {
 // correctly spaced by embedRateLimitInterval regardless of goroutine
 // scheduling order, rather than racing to read the same "last call time"
 // and under-pacing.
-func (r *embedRateLimiter) wait(ctx context.Context, ratePerSecond int) {
+func (r *embedRateLimiter) wait(ctx context.Context, ratePerSecond float64) {
 	interval := embedRateLimitInterval(ratePerSecond)
 	if interval <= 0 {
 		return
